types 40%

+++ inputs.pde
@@ -6,7 +6,7 @@ class Place {
   int x=0;
   int y=0;
   int wid=0;
-  int hei=0;
+  float hei=0;
   //int type=0;//nothing, button, switch, scroll
   int value=0;
   int barmax=1; //have to delete/make scroll only
